refactor(pages/api/auth): refactor session management and api errors sending

+++ pages/api/auth/verify-email.ts
@@ -1,8 +1,9 @@
 import { NextApiRequest, NextApiResponse } from 'next';
-import { getSession } from 'next-auth/react';
 import { tokenDataAccess, userDataAccess } from '../../../infrastructure/data-access';
 import { connectToDatabase } from '../../../infrastructure/database';
+import { getSessionUser } from '../../../services/auth/auth.api.service';
 import { sendAccountVerificationEmail } from '../../../utils/email.util';
+import { sendApiError } from '../../../utils/error.utils';
 import { generateToken, verifyToken } from '../../../utils/jwt.util';
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
@@ -11,40 +12,31 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
 
     if (req.method === 'GET') {
 
-        const session = await getSession({ req });
-        if (!session) {
-            return res.status(401).json({
-                code: 'auth/unauthorized',
-                message: 'Unauthorized.',
-            });
+        const currentUser = await getSessionUser(req);
+        if (!currentUser) {
+            return sendApiError(res, 'auth', 'unauthorized');
         }
 
-        const user = await userDataAccess.findUserById(session.user._id);
+        const userData = await userDataAccess.findUserById(currentUser._id);
 
-        if (!user) {
-            return res.status(404).json({
-                code: 'auth/user-not-found',
-                message: 'User not found.',
-            });
+        if (!userData) {
+            return sendApiError(res, 'auth', 'user-not-found');
         }
 
-        if (user.email_verified) {
-            return res.status(409).json({
-                code: 'auth/user-already-verified',
-                message: 'User email already verified.',
-            });
+        if (userData.email_verified) {
+            return sendApiError(res, 'auth', 'user-already-verified');
         }
 
         const expirationDate = Math.floor(Date.now() / 1000) + (60 * 60 * 24);
-        const token = generateToken(user, expirationDate, 'account_verification');
+        const token = generateToken(userData, expirationDate, 'account_verification');
         const savedToken = await tokenDataAccess.createToken({
             token,
             expiration_date: new Date(expirationDate),
             action: 'account_verification',
-            created_by: session.user._id,
+            created_by: currentUser._id,
         });
 
-        const emailResponse = sendAccountVerificationEmail(user, savedToken);
+        const emailResponse = sendAccountVerificationEmail(userData, savedToken);
 
         return res.status(200).json(emailResponse);
 
@@ -53,65 +45,45 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
     if (req.method === 'PUT') {
         const { token } = req.body;
 
-        const session = await getSession({ req });
-        if (!session) {
-            return res.status(401).json({
-                code: 'auth/unauthorized',
-                message: 'Unauthorized.',
-            });
+        const currentUser = await getSessionUser(req);
+
+        if (!currentUser) {
+            return sendApiError(res, 'auth', 'unauthorized');
         }
 
         if (!token) {
-            return res.status(401).json({
-                code: 'auth/invalid-token',
-                message: 'Invalid token.',
-            });
+            return sendApiError(res, 'auth', 'invalid-token');
         }
 
         const savedToken = await tokenDataAccess.getTokenFromTokenString(token);
 
         if (!savedToken) {
-            return res.status(404).json({
-                code: 'auth/token-not-found',
-                message: 'Token not found.',
-            });
+            return sendApiError(res, 'auth', 'token-not-found');
         }
 
         const tokenPayload = verifyToken(savedToken.token);
-        const user = await userDataAccess.findUserByEmail(tokenPayload.email);
+        const userData = await userDataAccess.findUserByEmail(tokenPayload.email);
         await tokenDataAccess.deleteTokenById(savedToken._id);
 
-        if (!user) {
-            return res.status(404).json({
-                code: 'auth/user-not-found',
-                message: 'User not found.',
-            });
+        if (!userData) {
+            return sendApiError(res, 'auth', 'user-not-found');
         }
 
-        if (typeof user._id === 'string' && user._id !== session.user._id || typeof user._id !== 'string' && user._id.toHexString() !== session.user._id) {
-            return res.status(401).json({
-                code: 'auth/wrong-token',
-                message: 'Provided token does not match the user.',
-            });
+        if (typeof userData._id === 'string' && userData._id !== currentUser._id || typeof userData._id !== 'string' && userData._id.toHexString() !== currentUser._id) {
+            return sendApiError(res, 'auth', 'wrong-token');
         }
 
-        if (user.email_verified) {
-            return res.status(409).json({
-                code: 'auth/user-already-verified',
-                message: 'User email already verified.',
-            });
+        if (userData.email_verified) {
+            return sendApiError(res, 'auth', 'user-already-verified');
         }
 
         const updatedUser = await userDataAccess.updateUser({
-            _id: user._id,
+            _id: userData._id,
             email_verified: true,
         });
 
         return res.status(200).json(updatedUser);
     }
 
-    res.status(405).json({
-        code: 'auth/wrong-method',
-        message: 'This request method is not allowed.',
-    });
+    return sendApiError(res, 'auth', 'wrong-method');
 }
